requirements.txt
alembic==1.11.1
asyncpg==0.27.0
bcrypt==4.0.1
fastapi==0.98.0
fastapi-restful==0.4.5
passlib==1.7.4
pre-commit==3.3.3
email-validator==2.0.0.post2
python-dotenv==1.0.0
python-jose==3.3.0
sqlalchemy==2.0.17
uvicorn==0.22.0
